fix copying RGP captures by checking for completion
Waiting 0.1s before starting the copy over SSH isn't enough because
the trace might not be completely written on disk and this can copy
corrupted captures.

To fix this, wait for completion by checking the RGP capture size
over SSH. This is simple and should be robust enough. An alternative
solution would be inotify but inotifywait isn't installed on Deck.

Closes: https://gitlab.steamos.cloud/devkit/steamos-devkit/-/issues/3